requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic-settings==2.5.0
stripe>=8.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
httpx>=0.27.0
